fix: Makefile lint 范围修复 + mypy 类型错误修复
- Makefile: mypy 添加 framework 目录检查
- algorithm_core.py: 修复 Collection[str] 索引赋值类型错误
- dcf.py: numpy floating 转 float 类型修复
- bayesian.py: max() key 参数类型修复

Tests: 1205 passed, coverage 81.90%

+++ framework/trading/seasons/dcf.py
@@ -311,13 +311,13 @@ class DCFValuation:
             valuations[i] = enterprise_value / shares_outstanding
 
         # 计算统计量
-        mean_val = np.mean(valuations)
-        median_val = np.median(valuations)
-        std_val = np.std(valuations)
+        mean_val = float(np.mean(valuations))
+        median_val = float(np.median(valuations))
+        std_val = float(np.std(valuations))
 
         # 置信区间
-        ci_90 = (np.percentile(valuations, 5), np.percentile(valuations, 95))
-        ci_95 = (np.percentile(valuations, 2.5), np.percentile(valuations, 97.5))
+        ci_90 = (float(np.percentile(valuations, 5)), float(np.percentile(valuations, 95)))
+        ci_95 = (float(np.percentile(valuations, 2.5)), float(np.percentile(valuations, 97.5)))
 
         # 发送事件
         Events.dcf_calculated.send(
